+++ internal/cmd/dtypes_codegen/enums.go
@@ -180,6 +180,6 @@ func generateEnums(contents string) {
 	f := must.M1(os.Create(DTypeEnumGoFileName))
 	must.M(enumsFromCTemplate.Execute(f, allValues))
 	must.M(exec.Command("gofmt", "-w", DTypeEnumGoFileName).Run())
-	fmt.Printf("Generated %q based on pjrt_c_api.h\n", DTypeEnumGoFileName)
+	fmt.Printf("✅ Successfully generated %q based on pjrt_c_api.h\n", DTypeEnumGoFileName)
 	must.M(exec.Command("enumer", "-type=DType", "-yaml", "-json", "-text", "-values", DTypeEnumGoFileName).Run())
 }