+++ pjrt/gen_api_calls.c
@@ -148,6 +148,17 @@ PJRT_Error* call_PJRT_Client_LookupAddressableDevice(const PJRT_Api *api, PJRT_C
 	return api->PJRT_Client_LookupAddressableDevice(args);
 }
 
+// call_PJRT_Client_UpdateGlobalProcessInfo calls the corresponding PJRT API method.
+// Updates the PjRt client with information about all global processes.
+//
+// Recall that a distributed program may consist of multiple PjRt clients
+// spanning multiple machines. These clients perform collective operations, like
+// AllGather, to execute a distributed program. UpdateGlobalProcessInfo updates
+// a PjRt client with information about all processes.
+PJRT_Error* call_PJRT_Client_UpdateGlobalProcessInfo(const PJRT_Api *api, PJRT_Client_UpdateGlobalProcessInfo_Args* args) {
+	return api->PJRT_Client_UpdateGlobalProcessInfo(args);
+}
+
 // call_PJRT_Client_AddressableMemories calls the corresponding PJRT API method.
 // Returns a list of memories that are addressable from the client. Addressable
 // memories are those that the client can directly transfer data to and from.
@@ -217,6 +228,11 @@ PJRT_Error* call_PJRT_AsyncHostToDeviceTransferManager_SetBufferError(const PJRT
 // call_PJRT_AsyncHostToDeviceTransferManager_AddMetadata calls the corresponding PJRT API method.
 PJRT_Error* call_PJRT_AsyncHostToDeviceTransferManager_AddMetadata(const PJRT_Api *api, PJRT_AsyncHostToDeviceTransferManager_AddMetadata_Args* args) {
 	return api->PJRT_AsyncHostToDeviceTransferManager_AddMetadata(args);
+}
+
+// call_PJRT_Client_CreateUninitializedBuffer calls the corresponding PJRT API method.
+PJRT_Error* call_PJRT_Client_CreateUninitializedBuffer(const PJRT_Api *api, PJRT_Client_CreateUninitializedBuffer_Args* args) {
+	return api->PJRT_Client_CreateUninitializedBuffer(args);
 }
 
 // call_PJRT_Client_BufferFromHostBuffer calls the corresponding PJRT API method.
@@ -788,6 +804,11 @@ PJRT_Error* call_PJRT_TopologyDescription_GetDeviceDescriptions(const PJRT_Api *
 // Serializes the TopologyDescription to a string for use in cache keys.
 PJRT_Error* call_PJRT_TopologyDescription_Serialize(const PJRT_Api *api, PJRT_TopologyDescription_Serialize_Args* args) {
 	return api->PJRT_TopologyDescription_Serialize(args);
+}
+
+// call_PJRT_TopologyDescription_Deserialize calls the corresponding PJRT API method.
+PJRT_Error* call_PJRT_TopologyDescription_Deserialize(const PJRT_Api *api, PJRT_TopologyDescription_Deserialize_Args* args) {
+	return api->PJRT_TopologyDescription_Deserialize(args);
 }
 
 // call_PJRT_TopologyDescription_Attributes calls the corresponding PJRT API method.