+++ xlabuilder/convolve.go
@@ -4,59 +4,69 @@ import "github.com/pkg/errors"
 
 // ConvolveAxesConfig defines the interpretation of the input/kernel/output tensor axes.
 // There must be the same number of spatial dimensions (axes) for each of the 3 tensors.
-// Input and output has batch and channel axes. Kernel has inputChannel and outputChannel axes.
+// The input and output have batch and channels axes.
+// The Kernel has an input feature and output channels axes.
+//
+// Note another common term for "channels" is "features".
 type ConvolveAxesConfig struct {
-	InputBatch, InputChannel int
-	InputSpatial             []int
+	InputBatch, InputChannels int
+	InputSpatial              []int
 
-	KernelInputChannel, KernelOutputChannel int
-	KernelSpatial                           []int
+	KernelInputChannels, KernelOutputChannels int
+	KernelSpatial                             []int
 
-	OutputBatch, OutputChannel int
-	OutputSpatial              []int
+	OutputBatch, OutputChannels int
+	OutputSpatial               []int
 }
 
-// ConvGeneralDilated is a generic Convolution operation offered by XLA.
-// featureAxisAfter defines whether the features (aka. channels or depth) axis comes after the
-// spatial dimension. Example: a 2D input can be one of the two:
+// ConvGeneral is a generic Convolution operation with support for:
 //
-//   - featureAxisAfter=false: input=[batch_size, features, height, width], filter=[output_features, input_features, height, width]
-//   - featureAxisAfter=true:  input=[batch_size, height, width, features], filter=[output_features, height, width, input_features]
+// - Arbitrary number of spatial axes.
+// - Arbitrary transposition of axes.
+// - Strides and padding.
+// - Dilations of the input.
+// - Dilations of the kernel, aka. atrous convolution.
+// - Channels grouping (on the input channels).
+// - Batch grouping.
 //
 // Some details in https://www.tensorflow.org/xla/operation_semantics#convwithgeneralpadding_convolution.
 // There operand and filter are called lhs and rhs.
 // (XLA documentation is unfortunately poor, much is guess-work).
 // Also useful, https://arxiv.org/pdf/1603.07285v1.pdf.
-func ConvGeneralDilated(operand, filter *Op, axes ConvolveAxesConfig,
-	strides []int, paddings [][2]int, inputDilation, filterDilation []int,
-	filterGroupCount, batchGroupCount int) (*Op, error) {
-	builder := operand.builder
-	op := newOp(ConvGeneralDilatedOp, operand, filter)
-	numSpatialDims := operand.Shape.Rank() - 2
+//
+// Note:
+//   - Another common term for "channels" is "features".
+//   - "Kernel" is also commonly called "weights" or "filters".
+func ConvGeneral(input, kernel *Op, axes ConvolveAxesConfig,
+	strides []int, paddings [][2]int, inputDilations, kernelDilations []int,
+	channelGroupCount, batchGroupCount int) (*Op, error) {
+	builder := input.builder
+	op := newOp(ConvGeneralDilatedOp, input, kernel)
+	numSpatialDims := input.Shape.Rank() - 2
 	if len(axes.InputSpatial) != numSpatialDims || len(axes.OutputSpatial) != numSpatialDims || len(axes.KernelSpatial) != numSpatialDims {
-		return nil, errors.Errorf("ConvGeneralDilated: operand has %d spatial dimensions, but axes configuration has InputSpatial=%d, KernelSpatial=%d, OutputSpatial=%d spatial axes configured "+
-			"for operand/kernel/output", numSpatialDims, len(axes.InputSpatial), len(axes.KernelSpatial), len(axes.OutputSpatial))
+		return nil, errors.Errorf("ConvGeneralDilated: input has %d spatial dimensions, but axes configuration has InputSpatial=%d, KernelSpatial=%d, OutputSpatial=%d spatial axes configured "+
+			"for input/kernel/output", numSpatialDims, len(axes.InputSpatial), len(axes.KernelSpatial), len(axes.OutputSpatial))
 	}
 
 	// Encoding of the values as follows. IMPORTANT: this code needs to be in sync with corresponding
 	// decoding code in c/gomlx/xlabuilder/xlabuilder.cpp, in function XlaBuilderAddOp, under ConvGeneralDilatedOp case.
 	//  * 8 first elements store the various parameters and lengths:
 	axesConfigLen := 3 * (numSpatialDims + 2)
-	op.IntsArg = make([]int, 0, 7+axesConfigLen+len(strides)+2*len(paddings)+len(inputDilation)+len(filterDilation))
+	op.IntsArg = make([]int, 0, 7+axesConfigLen+len(strides)+2*len(paddings)+len(inputDilations)+len(kernelDilations))
 	encode := func(values ...int) {
 		op.IntsArg = append(op.IntsArg, values...)
 	}
 
 	// Encode dimensions.
-	encode(numSpatialDims, filterGroupCount, batchGroupCount)
-	encode(len(strides), len(paddings), len(inputDilation), len(filterDilation))
+	encode(numSpatialDims, channelGroupCount, batchGroupCount)
+	encode(len(strides), len(paddings), len(inputDilations), len(kernelDilations))
 
 	// Append axes configuration.
-	encode(axes.InputBatch, axes.InputChannel)
+	encode(axes.InputBatch, axes.InputChannels)
 	encode(axes.InputSpatial...)
-	encode(axes.KernelInputChannel, axes.KernelOutputChannel)
+	encode(axes.KernelInputChannels, axes.KernelOutputChannels)
 	encode(axes.KernelSpatial...)
-	encode(axes.OutputBatch, axes.OutputChannel)
+	encode(axes.OutputBatch, axes.OutputChannels)
 	encode(axes.OutputSpatial...)
 
 	// Append arrays of ints.
@@ -64,8 +74,8 @@ func ConvGeneralDilated(operand, filter *Op, axes ConvolveAxesConfig,
 	for _, pair := range paddings {
 		encode(pair[0], pair[1])
 	}
-	encode(inputDilation...)
-	encode(filterDilation...)
+	encode(inputDilations...)
+	encode(kernelDilations...)
 
 	err := builder.addOp(op)
 	if err != nil {
@@ -74,12 +84,12 @@ func ConvGeneralDilated(operand, filter *Op, axes ConvolveAxesConfig,
 	return op, nil
 }
 
-// DecodeConvGeneralDilated retrieves the arguments for the ConvGeneralDilated op.
-func DecodeConvGeneralDilated(op *Op) (operand, filter *Op, axes ConvolveAxesConfig,
-	strides []int, paddings [][2]int, inputDilation, filterDilation []int,
-	filterGroupCount, batchGroupCount int) {
-	operand = op.OpInputs[0]
-	filter = op.OpInputs[1]
+// DecodeConvGeneral retrieves the arguments for the ConvGeneral op.
+func DecodeConvGeneral(op *Op) (input, kernel *Op, axes ConvolveAxesConfig,
+	strides []int, paddings [][2]int, inputDilations, kernelDilations []int,
+	channelGroupCount, batchGroupCount int) {
+	input = op.OpInputs[0]
+	kernel = op.OpInputs[1]
 	pos := 0
 	decode := func() int {
 		res := op.IntsArg[pos]
@@ -97,7 +107,7 @@ func DecodeConvGeneralDilated(op *Op) (operand, filter *Op, axes ConvolveAxesCon
 
 	// Encode dimensions.
 	numSpatialDims := decode()
-	filterGroupCount = decode()
+	channelGroupCount = decode()
 	batchGroupCount = decode()
 
 	stridesLen := decode()
@@ -107,13 +117,13 @@ func DecodeConvGeneralDilated(op *Op) (operand, filter *Op, axes ConvolveAxesCon
 
 	// Append axes configuration.
 	axes.InputBatch = decode()
-	axes.InputChannel = decode()
+	axes.InputChannels = decode()
 	axes.InputSpatial = decodeN(numSpatialDims)
-	axes.KernelInputChannel = decode()
-	axes.KernelOutputChannel = decode()
+	axes.KernelInputChannels = decode()
+	axes.KernelOutputChannels = decode()
 	axes.KernelSpatial = decodeN(numSpatialDims)
 	axes.OutputBatch = decode()
-	axes.OutputChannel = decode()
+	axes.OutputChannels = decode()
 	axes.OutputSpatial = decodeN(numSpatialDims)
 
 	// Append arrays of ints.
@@ -125,7 +135,7 @@ func DecodeConvGeneralDilated(op *Op) (operand, filter *Op, axes ConvolveAxesCon
 			paddings[ii][1] = decode()
 		}
 	}
-	inputDilation = decodeN(inputDilationLen)
-	filterDilation = decodeN(filterDilationLen)
+	inputDilations = decodeN(inputDilationLen)
+	kernelDilations = decodeN(filterDilationLen)
 	return
 }